vehicle_damage_detection/requirements.txt
tensorflow>=2.12.0
tensorflow-addons>=0.19.0
numpy>=1.23.0
opencv-python>=4.7.0
albumentations>=1.3.0
pycocotools>=2.0.6
Pillow>=9.4.0
scikit-learn>=1.2.0
scipy>=1.10.0
matplotlib>=3.6.0
pyyaml>=6.0
tqdm>=4.64.0
wandb>=0.13.0
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
python-multipart>=0.0.6
pydantic>=1.10.0
aiofiles>=23.1.0
requests>=2.28.0
pytest>=7.2.0
pytest-asyncio>=0.21.0
httpx>=0.23.0
